+++ src/main/query.js
@@ -3,10 +3,12 @@ import { gql } from 'apollo-boost';
 export const LOOP_QUERY = gql`
   query LoopQuery {
     allSettings {
-      pageForPosts
-      pageOnFront
-      permalinkStructure
+      readingSettingsPageForPosts
+      readingSettingsPageOnFront
       readingSettingsPostsPerPage
+      permalinkSettingsStructure
+      permalinkSettingsTagBase
+      permalinkSettingsCategoryBase
     }
   }
 `;